Makefile; start/stop/status commands
New Makefile to build the commands and (re)start the server.

Added the code to properly return the status of the crawl as the
enum CrawlStatus; previously, this was always coming back as STOPPED.

Ready to hook up the actual crawlers to the server.
TODO: real Fetcher class; real crawler stop/start