Shared tree package
Implements a shared gotree.

Add requests are synchronous, and return a pointer into the tree
that can be used in subsequent calls as the new insert point. Format
requests dump the current state of the tree using gotree's Print()
method. Finally, sending a bool to Quit shuts down the process.